Apply CoR pattern
